Show thumbnails in image gallery block
The block is now successfully handling caching. It is detecting whether
it's on a product page or on a page the references a product. It shows
the thumbnails of any product images that have been uploaded for the
product.

Next up is to add the modal functionality to display a larger image.