Save test work for POSTing via WEA custom resource
I wanted to ensure that having two separate resources was going to work
once it came time to do the POST verb. I was concerned that having a
resource with a canonical link of /wea/actions and another resource
that used a "canonical" link of /wea/actions/{id} and the
"https://www.drupal.org/link-relations/create" link of /wea/actions
was going to conflict. As it turns out, that does not appear to be a
problem!