Save Language Manager logic
I don't think I'm going to need the logic because I'm typehinting with
Node for the get method. But, I want to have all this code somewhere in
case I do end up needing to retrieve it. I haven't done any testing with
it so I don't know whether it works or not.