Add a .install / hook_install
There is a bug (https://www.drupal.org/node/2599228) that deals with
translatable content types that are defined in code causing an error
about missing the content_translation_source column when saving content.
One workaround was to create a module_name.install file that implements
the hook_install hook to do

<code>
  \Drupal::service('entity.definition_update_manager')->applyUpdates();
</code>

Applying this workaround does fix the issue I was having with my
Functional test.