Save the debugging work I did
The Functional test is creating a product piece of content that has an
image attached. Unfortunately when it visits the Controller route to
display the image for that piece of content it is returning the message
that the piece of content does not have an image at the delta.

I wanted to ensure that the Functional test was creating the product
piece of content as I expected. So, I created a test Controller to
replicate the same thing (more or less) that is happening in the
Functional test.

I needed to modify the routing file so that I could access the
Controller.

I have a couple bits of logic in the Controller.
* There's the bit that created the node intially.
* And there's the bit that I added so that I could keep on calling the
  Controller without continuing to change things and so that I could have
  a look at what the node looks like behind the scenes.

The result of this testing is that the product piece of content is being
created exactly as I expect it should be. So, the issue is not in how
the Functional test is creating the piece of content. I need to figure
out why the Controller can't see that there is an image attached to that
piece of content.