Add a comment
I'm going to need to go through all of the code to see where I can /
should add comments so that students can learn from them. I want to avoid
going through the code line-by-line in the videos; I'd rather proceed
quickly and then direct the students to the comments.