Switch to using the entity type manager
I had initially injected the node storage directly and have now decided
to switch to injeting the entity type manager and then using the
getStorage method in the constructor to retrieve the node storage. I'm
doing this because I think it will make it easier for the students when
they are completing the assignment to create the block. I'll be giving them
the hint that they need to retrieve the node storage service in their
constructor.